Move Combine helpers to a new file.
